backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "pymongo>=4.14.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "google-genai>=0.1.0",
    "motor>=3.4.0",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "lxml>=4.9.0",
    "readability-lxml>=0.8.1",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.20",
    "certifi>=2025.8.3",
    "jwt>=1.4.0",
]
